requirements.txt
aiohttp==1.0.3
async-timeout==1.0.0
cffi==1.8.3
chardet==2.3.0
command-not-found==0.3
configparser==3.5.0
discord.py==0.13.0
google-api-python-client==1.5.3
httplib2==0.9.2
language-selector==0.1
multidict==2.1.2
oauth2client==3.0.0
peewee==2.8.5
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pycurl==7.19.3
pygobject==3.12.0
PyNaCl==1.0.1
python-apt===0.9.3.5ubuntu1
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
ufw===0.34-rc-0ubuntu2
unattended-upgrades==0.1
uritemplate==0.6
websockets==3.2
youtube-dl==2016.10.7
